Moved command files to their proper directory. Changed DiskMark to public.
