+++ lib/src/main/java/edu/touro/mco152/bm/DiskWorker.java
@@ -1,24 +1,14 @@
 package edu.touro.mco152.bm;
 
 import edu.touro.mco152.bm.commands.ICommand;
+import edu.touro.mco152.bm.commands.readBM;
 import edu.touro.mco152.bm.commands.simpleInvoker;
-import edu.touro.mco152.bm.persist.DiskRun;
-import edu.touro.mco152.bm.persist.EM;
+import edu.touro.mco152.bm.commands.writeBM;
 import edu.touro.mco152.bm.ui.Gui;
 
-import jakarta.persistence.EntityManager;
 import javax.swing.*;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.RandomAccessFile;
-import java.util.Date;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 import static edu.touro.mco152.bm.App.*;
-import static edu.touro.mco152.bm.DiskMark.MarkType.READ;
-import static edu.touro.mco152.bm.DiskMark.MarkType.WRITE;
 
 /**
  * Run the disk benchmarking as a Swing-compliant thread (only one of these threads can run at
