+++ lib/src/main/java/edu/touro/mco152/bm/commands/writeBM.java
@@ -1,5 +1,9 @@
-package edu.touro.mco152.bm;
+package edu.touro.mco152.bm.commands;
 
+import edu.touro.mco152.bm.App;
+import edu.touro.mco152.bm.DiskMark;
+import edu.touro.mco152.bm.IOutput;
+import edu.touro.mco152.bm.Util;
 import edu.touro.mco152.bm.commands.ICommand;
 import edu.touro.mco152.bm.persist.DiskRun;
 import edu.touro.mco152.bm.persist.EM;
